changed float to double

+++ src/lj_lifeos.c
@@ -442,14 +442,14 @@ static void unloadPromoter() {
 }
 
 static int lua_delay(lua_State *L) {
-	float secs = luaL_optnumber(L, 1, 0);
-    sceKernelDelayThread(secs * 1000*1000); // this converts microsecs to secs
+	double secs = luaL_optnumber(L, 1, 0);
+    sceKernelDelayThread(secs * 1000*1000);
     return 0;
 }
 
 static int lua_delaycb(lua_State *L) {
-	float secs = luaL_optnumber(L, 1, 0);
-    sceKernelDelayThreadCB(secs * 1000*1000); // this converts microsecs to secs
+	double secs = luaL_optnumber(L, 1, 0);
+    sceKernelDelayThreadCB(secs * 1000*1000);
     return 0;
 }
